+++ src/gametest/java/com/warlonmhite/hempdustry/test/HempdustryGameTests.java
@@ -50,6 +50,7 @@ public class HempdustryGameTests implements ModInitializer {
         register("a_full_green_out_ends_the_high", RebalanceGameTest::aFullGreenOutEndsTheHigh);
         register("edibles_buy_time_not_level", RebalanceGameTest::ediblesBuyTimeNotLevel);
         register("a_broken_config_is_left_alone", ConfigGameTest::aBrokenConfigIsLeftAlone);
+        register("a_hemp_boat_does_not_stop_the_data_fixer", DataFixerGameTest::aHempBoatDoesNotStopTheDataFixer);
     }
 
     private static void register(String name, Consumer<TestContext> test) {